requirements.txt
asgiref==3.8.1
Django==5.2
pillow==11.1.0
pygame==2.6.1
sqlparse==0.5.3
django-ninja>=1.5.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.6.0
django-ninja-jwt>=5.4.0
